removing need for implicits and injected ldap helper, while keeping better error messaging
