+++ app/controllers/System.scala
@@ -2,7 +2,7 @@ package controllers
 
 import javax.inject.Inject
 
-import auth.{LDAPConnectionPoolWrapper, Security}
+import auth.Security
 import com.unboundid.ldap.sdk.LDAPConnectionPool
 import play.api.{Configuration, Logger}
 import play.api.db.slick.DatabaseConfigProvider
@@ -16,11 +16,10 @@ import helpers.DatabaseHelper
 import play.api.cache.SyncCacheApi
 
 class System @Inject() (cc:ControllerComponents, configuration: Configuration, dbConfigProvider: DatabaseConfigProvider,
-                        databaseHelper:DatabaseHelper,cacheImpl:SyncCacheApi, ldapPool:LDAPConnectionPoolWrapper)
+                        databaseHelper:DatabaseHelper,cacheImpl:SyncCacheApi)
   extends AbstractController(cc) with Security {
 
   implicit val cache:SyncCacheApi = cacheImpl
-  implicit val ldapConnectionPool:LDAPConnectionPool = ldapPool.connectionPool.getOrElse(null)
   private val dbConfig = dbConfigProvider.get[JdbcProfile]
   private val logger: Logger = Logger(this.getClass)
