+++ conf/application.conf
@@ -98,8 +98,6 @@ ldap {
   ldapPort = 636
   ldapHost0 = "adhost1.myorg.int"
   ldapHost1 = "adhost2.myorg.int"
-  serverAddresses = ["adhost1.myorg.int","adhost2.myorg.int"]
-  serverPorts = [ldapPort,ldapPort]
   bindDN = "aduser"
   bindPass = "adpassword"
   poolSize = 3
